+++ Scripts/Items/Addons/DawnsMusicBox/DawnsMusicGear.cs
@@ -87,7 +87,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.WriteEncodedInt((int)1); // version
+            writer.WriteEncodedInt(1); // version
 
             writer.Write((int)this.m_Music);
         }
