+++ Scripts/Items/Addons/DawnsMusicBox/RewardVersion/DawnsMusicBox.cs
@@ -256,7 +256,7 @@ namespace Server.Items.MusicBox
             {
                 TrackInfo ti = TrackInfo.GetInfo(this.m_ActualSong);
 
-                this.m_PlayingTimer = new PlayingTimer((double)ti.Duration, this);	// add a new timer
+                this.m_PlayingTimer = new PlayingTimer(ti.Duration, this);	// add a new timer
                 this.m_PlayingTimer.Start();
             }
         }
@@ -291,7 +291,7 @@ namespace Server.Items.MusicBox
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
 
             writer.Write(this.m_Tracks.Count);
