+++ Scripts/Items/Artifacts/Equipment/Spellbooks/Runebook.cs
@@ -145,7 +145,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)3);
+            writer.Write(3);
 
             writer.Write((byte)m_Quality);
