+++ Scripts/Items/Consumables/BasePotion.cs
@@ -158,7 +158,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)1); // version
+            writer.Write(1); // version
 
             writer.Write((int)m_PotionEffect);
         }
