+++ Scripts/Items/Consumables/FearEssence.cs
@@ -58,7 +58,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
@@ -99,7 +99,7 @@ namespace Server.Items
             {
                 if (mobile != from && from.CanBeHarmful(mobile, false))
                 {
-                    double chance = (double)(((4 * mobile.Skills[SkillName.MagicResist].Value) + 150) / 700);
+                    double chance = ((4 * mobile.Skills[SkillName.MagicResist].Value) + 150) / 700;
 
                     if (chance < Utility.RandomDouble())
                     {
